+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 0.2.0 (2023/07/19)
+
+- Fixed error handling
+- Added instant mode to disable typing simulation
+- Fixed typing simulation speed
+
 ## Version 0.1.0 (2023/07/19)
 
 - First version